Add solution for problem `D - Pashmak and Parmida's problem` Link: https://codeforces.com/contest/459/problem/D Tags: data structures, divide and conquer, sortings, *1800 Ref: https://codeforces.com/contest/459/submission/77203750

+++ codeforces/459/D.cpp
@@ -1,20 +1,20 @@
 #include<bits/stdc++.h>
 using namespace std;
 #define mx 2000000
-int fin[mx];
+long long fin[mx];
 
-void add(int p)
+void add(long long p)
 {
-for(int i=p;i<=mx;i+=(i&-i))
+for(long long i=p;i<=mx;i+=(i&-i))
 {
 fin[i]+=1;
 }
 
 }
-int sum(int p)
+int sum(long long p)
 {
-int ans=0;
-for(int i=p;i;i-=(i&-i))
+long long ans=0;
+for(long long i=p;i;i-=(i&-i))
 {
 ans+=fin[i];
 }
@@ -24,28 +24,28 @@ int main()
 {
    ios_base::sync_with_stdio(false);
     cin.tie(NULL);
-int n;
+long long n;
 cin>>n;
-map<int,int >ma;
-map<int,int >ma1;
-vector<int>sura;
-int ans1[n+1],ans2[n+1];
-for(int i=0;i<n;i++)
+map<long long,long long >ma;
+map<long long,long long >ma1;
+vector<long long>sura;
+long long ans1[n+1],ans2[n+1];
+for(long long i=0;i<n;i++)
 {
-int k;
+long long k;
 cin>>k;
 sura.push_back(k);
 ma[k]++;
 ans1[i]=ma[k];
 }
-for(int i=n-1;i>=0;i--)
+for(long long i=n-1;i>=0;i--)
 {
 
 ma1[sura[i]]++;
 ans2[i]=ma1[sura[i]];
 }
 long long ans=0;
-for(int i=n-1;i>=0;i--)
+for(long long i=n-1;i>=0;i--)
 {
 ans+=sum(ans1[i]);
 add(ans2[i]+1);
